Add adafruit-nrfutil
Something's been done to the adafruit version of the nrf52 bootloader
that makes it incompatiable with the normal nrf dfu utilities, which
means we can't use an open reimplementation.  In theory if we throw out
signing.py all the code in this is GPL as that is the only code that
marked with Nordic Simi's custom license, all the others are marked with
BSD.

Nordic Simi's license pretty much says you can only use it to build for
nordic simi chips, so no using this tool to flash a raspberry pi pico.

This tool provides two useful uses.
1. It converts the hex into a bin, creates a header, and a json file and
   then packages all that into a zip.
2. Connects to the bootloader via a serial port and pushes that binary
   data onto the microcontroller, thus flashing with no mounting and no
   borking of the kernel when the device just decides to disappear
   without warning.