Add flash script for NiceNano with DFU
This is an alternative flashing method for the Nice Nano V2's.  It will
work with other nrf52 devices with some extending.

The user only needs to put the n!n into bootloader mode and they will
able to be flashed without requiring additional steps.  As well, because
the system no-longer has a mounted filesystem just up and disappearing
there wont be left behind dead device references and the user doesn't
have to worry about umount getting locked up into uninterruptible sleep,
which means the user needs to reboot with a system that cannot cleanly
unmount the root filesystem anymore.