Add udev rules to support block-based flashing
The udev rules result in the n!n, and RP2 bootloaders showing up at
/dev/mapper/uf2_bootloader_${SERIAL} with global read/write permissions.
This will be used to support automated flashing without requiring
elevated privileges and without mounting a filesystem that will
disappear, which regularly upsets the kernel.

0666 permissions shouldn't be too big an issue because of how
short-lived the device will be there, and the user has to physically put
the device into bootloader mode, so presumably if something else flashes
it, they'll notice.

Additional bootloaders can be added by adding additional rules to the
udev rules between the `#### DEVICE BOOTLOADERS` and
`#### END DEVICE BOOTLOADERS` section.

I've included the Nice!Nano bootloader, which should encompass all nrf52
devices with the Adafruit bootloader as well as the RP2 bootloader,
which should cover the RPiPico, and the ProMicro-style variants of such.

Example usage of udev on NixOS:

{
  ...
  inputs.zmk-nix.url = "github:lilyinstarlight/zmk-nix";

  outputs = {
    zmk-nix,
    ...
  }: {
    nixosConfigurations = {
      myHost = nixosSystem {
        ...
        modules = [
          zmk-nix.nixosModules.udevRules
        ];
        ...
      };
    };
  };
}

For Non-NixOS use, the user should copy the uf2-block.rules file to
/etc/udev/rules.d/{priority}-uf2-block.rules.